+++ Backend/src/main/java/gwangju/ssafy/backend/domain/reservation/repository/ProductRepository.java
@@ -1,6 +1,7 @@
 package gwangju.ssafy.backend.domain.reservation.repository;
 
 import gwangju.ssafy.backend.domain.reservation.entity.Product;
+import gwangju.ssafy.backend.domain.reservation.entity.enums.ProductCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -12,5 +13,5 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
     List<Product> findAllByGroupId(Long groupId);
 
     // 카테고리에 따른 해당 그룹의 대여 물품들(예약 물품) 조회
-    List<Product> findAllByGroupIdAndCategory(Long groupId, String category);
+    List<Product> findAllByGroupIdAndCategory(Long groupId, ProductCategory category);
 }
