+++ Backend/src/main/java/gwangju/ssafy/backend/domain/reservation/service/ProductService.java
@@ -1,9 +1,12 @@
 package gwangju.ssafy.backend.domain.reservation.service;
 
 import gwangju.ssafy.backend.domain.reservation.dto.ProductSimpleInfo;
+import gwangju.ssafy.backend.domain.reservation.entity.enums.ProductCategory;
 
 import java.util.List;
 
 public interface ProductService {
-    List<ProductSimpleInfo> searchProductMyGroup(Long groupId);
+    List<ProductSimpleInfo> searchProduct(Long groupId);
+
+    List<ProductSimpleInfo> searchProductByCategory(Long groupId, ProductCategory category);
 }
