+++ Backend/src/main/java/gwangju/ssafy/backend/domain/reservation/service/impl/ProductServiceImpl.java
@@ -2,6 +2,7 @@ package gwangju.ssafy.backend.domain.reservation.service.impl;
 
 import gwangju.ssafy.backend.domain.reservation.dto.ProductSimpleInfo;
 import gwangju.ssafy.backend.domain.reservation.entity.Product;
+import gwangju.ssafy.backend.domain.reservation.entity.enums.ProductCategory;
 import gwangju.ssafy.backend.domain.reservation.repository.ProductRepository;
 import gwangju.ssafy.backend.domain.reservation.service.ProductService;
 import jakarta.transaction.Transactional;
@@ -20,8 +21,21 @@ public class ProductServiceImpl implements ProductService {
 
     // 해당 그룹의 대여 물품들 전체 조회
     @Override
-    public List<ProductSimpleInfo> searchProductMyGroup(Long groupId) {
+    public List<ProductSimpleInfo> searchProduct(Long groupId) {
         List<Product> productList = productRepository.findAllByGroupId(groupId);
+        List<ProductSimpleInfo> productSimpleInfoList = new ArrayList<>();
+        for(Product product: productList) {
+            ProductSimpleInfo productSimpleInfo = new ProductSimpleInfo();
+            productSimpleInfo.convert(product);
+            productSimpleInfoList.add(productSimpleInfo);
+        }
+        return productSimpleInfoList;
+    }
+
+    // 해당 그룹의 대여 물품들 카테고리 별 조회
+    @Override
+    public List<ProductSimpleInfo> searchProductByCategory(Long groupId, ProductCategory category) {
+        List<Product> productList = productRepository.findAllByGroupIdAndCategory(groupId, category);
         List<ProductSimpleInfo> productSimpleInfoList = new ArrayList<>();
         for(Product product: productList) {
             ProductSimpleInfo productSimpleInfo = new ProductSimpleInfo();
