+++ Backend/src/main/java/gwangju/ssafy/backend/domain/user/exception/UserException.java
@@ -1,25 +1,26 @@
-package gwangju.ssafy.backend.global.exception;
+package gwangju.ssafy.backend.domain.user.exception;
 
 import lombok.Getter;
 
 @Getter
 public class UserException extends RuntimeException {
-    private final ErrorCode errorCode;
+    private final UserError errorCode;
     private final int status;
     private final String errorMessage;
 
 
-    public UserException(ErrorCode errorCode) {
+    public UserException(UserError errorCode) {
         super(errorCode.getErrorMessage());
         this.errorCode = errorCode;
         this.status = errorCode.getHttpStatus().value();
         this.errorMessage = errorCode.getErrorMessage();
     }
 
-    public UserException(ErrorCode errorCode, Throwable e) {
+    public UserException(UserError errorCode, Throwable e) {
         super(errorCode.getErrorMessage(), e);
         this.errorCode =errorCode;
         this.status = errorCode.getHttpStatus().value();
         this.errorMessage = errorCode.getErrorMessage();
     }
+
 }
