+++ Backend/src/main/java/gwangju/ssafy/backend/global/exception/TokenException.java
@@ -4,19 +4,19 @@ import lombok.Getter;
 
 @Getter
 public class TokenException extends RuntimeException {
-    private final ErrorCode errorCode;
+    private final GlobalError errorCode;
     private final int status;
     private final String errorMessage;
 
 
-    public TokenException(ErrorCode errorCode) {
+    public TokenException(GlobalError errorCode) {
         super(errorCode.getErrorMessage());
         this.errorCode = errorCode;
         this.status = errorCode.getHttpStatus().value();
         this.errorMessage = errorCode.getErrorMessage();
     }
 
-    public TokenException(ErrorCode errorCode, Throwable e) {
+    public TokenException(GlobalError errorCode, Throwable e) {
         super(errorCode.getErrorMessage(), e);
         this.errorCode = errorCode;
         this.status = errorCode.getHttpStatus().value();
